requirements.txt
flask
flask-cors
flask-sqlalchemy
python-dotenv
requests
google-genai
gunicorn
PyJWT
werkzeug